crypt-keeper-django/requirements.txt
defusedxml==0.4.1
Django==1.8.15
django-tastypie==0.13.3
lxml==3.6.4
python-dateutil==2.5.3
python-mimeparse==1.5.3
PyYAML==3.12
six==1.10.0
